Rename strategy, bucketserver and volumeserver packages
