+++ internal/strategy/strategy.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023 SAP SE or an SAP affiliate company and IronCore contributors
 // SPDX-License-Identifier: Apache-2.0
 
-package snapshotstrategy
+package strategy
 
 import (
 	"crypto/rand"
